Lazify `hyperspy.misc.utils`
