+++ hyperspy/components1d.py
@@ -20,7 +20,7 @@ import importlib
 
 from hyperspy.extensions import EXTENSIONS as _EXTENSIONS
 
-__all__ = [component for component, specs_ in _EXTENSIONS["components1D"].items()]
+__all__ = [component for component, _ in _EXTENSIONS["components1D"].items()]
 
 
 def __dir__():
