+++ hyperspy/components2d.py
@@ -21,7 +21,7 @@ import importlib
 
 from hyperspy.extensions import EXTENSIONS as _EXTENSIONS
 
-__all__ = [component for component, specs_ in _EXTENSIONS["components2D"].items()]
+__all__ = [component for component, _ in _EXTENSIONS["components2D"].items()]
 
 
 def __dir__():
